refactor: adds solana info
Also renames SolanaWallet to SolanaAccount in accordance with Solana's terminology
